+++ src/ZilaWSCallback.ts
@@ -5,6 +5,4 @@
  * MIT License
  * Copyright (c) 2023 ZilaWS
  */
-import IZilaClient from "./ZilaClient";
-
-export type ZilaWSCallback = (socket: IZilaClient, ...args: any[]) => any;
+export type ZilaWSCallback<T> = (socket: T, ...args: any[]) => any;